+++ fractal_server/tasks/v2/ssh/__init__.py
@@ -1,0 +1,3 @@
+from .collect import collect_ssh  # noqa
+from .deactivate import deactivate_ssh  # noqa
+from .reactivate import reactivate_ssh  # noqa