+++ tests/data/example_server_startup/run_server_uvicorn.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# Create an empty db
-poetry run fractalctl set-db
-
-# Start the server
-poetry run fractalctl start --port 8000