+++ tests/v2/06_tasks/test_collect_local.py
@@ -1,8 +1,4 @@
-from typing import Optional
-
-import pytest
 from devtools import debug
-from pydantic import BaseModel
 
 from fractal_server.app.models.v2 import TaskGroupActivityV2
 from fractal_server.app.models.v2 import TaskGroupV2
@@ -10,58 +6,7 @@ from fractal_server.app.schemas.v2 import (
     TaskGroupActivityStatusV2,
 )
 from fractal_server.app.schemas.v2.task_group import TaskGroupActivityActionV2
-from fractal_server.tasks.v2.local.collect import collect_local
-from fractal_server.tasks.v2.ssh.collect import collect_ssh
-from fractal_server.tasks.v2.utils_background import (
-    check_task_files_exist,
-)
-from fractal_server.tasks.v2.utils_database import _get_task_type
-
-
-class _MockTaskCreateV2(BaseModel):
-    name: str = "task_name"
-    command_non_parallel: Optional[str] = None
-    command_parallel: Optional[str] = None
-
-
-def test_get_task_type():
-
-    task = _MockTaskCreateV2(command_non_parallel="x")
-    assert _get_task_type(task) == "non_parallel"
-    task = _MockTaskCreateV2(command_parallel="x")
-    assert _get_task_type(task) == "parallel"
-    task = _MockTaskCreateV2(command_parallel="x", command_non_parallel="y")
-    assert _get_task_type(task) == "compound"
-
-
-def test_check_task_files_exist(tmp_path):
-    existing_path = tmp_path / "existing.py"
-    missing_path = tmp_path / "missing.py"
-    existing_path.touch()
-    existing_path = existing_path.as_posix()
-    missing_path = missing_path.as_posix()
-    # Success
-    check_task_files_exist(
-        task_list=[
-            _MockTaskCreateV2(command_non_parallel=f"py {existing_path}"),
-            _MockTaskCreateV2(command_parallel=f"py {existing_path}"),
-        ]
-    )
-    # Failures
-    with pytest.raises(FileNotFoundError) as e:
-        check_task_files_exist(
-            task_list=[
-                _MockTaskCreateV2(command_non_parallel=f"py {missing_path}")
-            ]
-        )
-    assert "missing file" in str(e.value)
-    with pytest.raises(FileNotFoundError) as e:
-        check_task_files_exist(
-            task_list=[
-                _MockTaskCreateV2(command_parallel=f"py {missing_path}")
-            ]
-        )
-    assert "missing file" in str(e.value)
+from fractal_server.tasks.v2.local import collect_local
 
 
 async def test_collect_pip_existing_file(tmp_path, db, first_user):
@@ -115,7 +60,6 @@ async def test_collect_pip_local_fail_rmtree(
     current_py_version,
     monkeypatch,
 ):
-
     import fractal_server.tasks.v2.local.collect
 
     def patched_function(*args, **kwargs):
@@ -182,26 +126,3 @@ async def test_collect_pip_local_fail_rmtree(
     assert task_group_activity_v2.status == "failed"
     assert "Broken rm" in task_group_activity_v2.log
     assert path.exists()
-
-
-def test_unit_missing_objects(db, caplog):
-    """
-    Test a branch which is in principle unreachable.
-    """
-    caplog.clear()
-    collect_local(
-        task_group_activity_id=9999,
-        task_group_id=9999,
-    )
-    assert "Cannot find database rows" in caplog.text
-
-    caplog.clear()
-    assert caplog.text == ""
-
-    collect_ssh(
-        task_group_activity_id=9999,
-        task_group_id=9999,
-        fractal_ssh=None,
-        tasks_base_dir="/invalid",
-    )
-    assert "Cannot find database rows" in caplog.text